move breadcrumb to components and make regex changes to breadcrumb
